Extract dashboard logic; enhance PDF export
Introduce a new useDashboardData hook that centralizes dashboard data fetching, processing and state (reservation/billing metrics, chart and ranking data, unit counts and active reservations). Refactor DashboardPage to consume the hook (simpler component, removed inlined data-fetch logic) and wire up refresh/filter behavior. Update ReportExportButtons: add ReservationOverview types and accept activeReservations prop; add a detailed reservations register page to PDF export with summary cards, paginated reservation rows, improved text wrapping for long descriptions, and consistent per-page footer. Misc: minor UI/comment cleanup in DashboardPage.
